+++ feature-libs/customer-ticketing/assets/translations/en/my-account-v2-customer-ticketing.ts
@@ -1,0 +1,8 @@
+export const myAccountV2CustomerTicketing = {
+  myAccountV2CustomerTicketing: {
+    heading: 'Customer Service',
+    showMore: 'Show More',
+    changedOn: 'Changed On: {{value}}',
+    ticketId: 'ID: {{value}}',
+  },
+};
