+++ feature-libs/customer-ticketing/components/my-account-v2/my-account-v2-customer-ticketing.component.ts
@@ -12,11 +12,11 @@ import {
 import { Observable } from 'rxjs';
 
 @Component({
-  selector: 'cx-myaccount-v2-customer-ticketing',
-  templateUrl: './myaccount-v2-customer-ticketing.component.html',
+  selector: 'cx-my-account-v2-customer-ticketing',
+  templateUrl: './my-account-v2-customer-ticketing.component.html',
 })
 export class MyAccountV2CustomerTicketingComponent {
-  protected PAGE_SIZE = 1;
+  protected readonly PAGE_SIZE = 1;
   protected customerTicketingFacade = inject(CustomerTicketingFacade);
   tickets$: Observable<TicketList | undefined> =
     this.customerTicketingFacade.getTickets(this.PAGE_SIZE);
