+++ feature-libs/customer-ticketing/components/public_api.ts
@@ -8,4 +8,4 @@ export * from './customer-ticketing-components.module';
 export * from './details/index';
 export * from './list/index';
 export * from './shared/index';
-export * from './myaccount-v2/index';
+export * from './my-account-v2/index';
