+++ feature-libs/customer-ticketing/styles/components/_index.scss
@@ -2,13 +2,13 @@
 @import './customer-ticketing-details';
 @import './customer-ticketing-dialog';
 @import './customer-ticketing-list';
-@import './myaccount-v2-customer-ticketing';
+@import './my-account-v2-customer-ticketing';
 
 $customer-ticketing-allowlist: cx-customer-ticketing-create-dialog,
   cx-customer-ticketing-reopen-dialog, cx-customer-ticketing-close-dialog,
   cx-customer-ticketing-reopen, cx-customer-ticketing-close,
   cx-customer-ticketing-create, cx-customer-ticketing-details,
-  cx-customer-ticketing-list, cx-myaccount-v2-customer-ticketing !default;
+  cx-customer-ticketing-list, cx-my-account-v2-customer-ticketing !default;
 
 $skipComponentStyles: () !default;
