+++ feature-libs/customer-ticketing/styles/components/_my-account-v2-customer-ticketing.scss
@@ -1,4 +1,4 @@
-%cx-myaccount-v2-customer-ticketing {
+%cx-my-account-v2-customer-ticketing {
   margin: 1rem 0rem;
   border: 1px solid var(--cx-color-medium);
   .cx-myaccount-customer-ticket-header {
