+++ feature-libs/order/assets/translations/en/index.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { MyAccountV2Order, order } from './order.i18n';
+import { MyAccountV2Order } from './my-account-v2-order.i18n';
+import { order } from './order.i18n';
 
 export const en = {
   order,
