+++ feature-libs/order/components/my-account-v2/index.ts
@@ -4,5 +4,5 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export * from './myaccount-v2-orders.component';
-export * from './myaccount-v2-orders.module';
+export * from './my-account-v2-orders.component';
+export * from './my-account-v2-orders.module';
