+++ feature-libs/order/styles/components/_index.scss
@@ -2,4 +2,4 @@
 @import './order-overview/index';
 @import './order-details/index';
 @import './replenishment-order/index';
-@import './myaccount-v2/index';
+@import './my-account-v2/index';
