+++ feature-libs/order/styles/components/my-account-v2/_index.scss
@@ -1,6 +1,6 @@
-@import './myaccount-v2-order';
+@import './my-account-v2-order';
 
-$myaccount-view-order-allowlist: cx-myaccount-v2-orders !default;
+$myaccount-view-order-allowlist: cx-my-account-v2-orders !default;
 
 $skipComponentStyles: () !default;
