+++ feature-libs/order/styles/components/my-account-v2/_my-account-v2-order.scss
@@ -1,4 +1,4 @@
-%cx-myaccount-v2-orders {
+%cx-my-account-v2-orders {
   margin: 1rem 0rem;
   border: 1px solid var(--cx-color-medium);
   .cx-myaccount-view-header {
