+++ projects/storefrontstyles/scss/components/content/navigation/_my-account-v2-navigation.scss
@@ -40,7 +40,7 @@
       }
     }
   }
-//to apply style to the icon from the stleClasses
+  //to apply style to the icon from the stleClasses
   .fas {
     font-family: sans-serif, 'Font Awesome 5 Free';
     font-size: var(--cx-font-medium);
